Remove CRM_Contact_Form_Search_Custom_Basic from searches added on install.
From https://github.com/civicrm/civicrm-core/pull/15963 this search appears to do what the main basic search does but
we can see that to make it work with sort etc we have to make changes to the BAO_Query object - doing
that to support a custom search scares me & we have a big picture position that we should move these
from core to an extension. However, an easy first step is not to add on new installs. In the case of
this search it seems like an easy call as it doesn't add anything discernable