Add caching to deprecation notice.
We have been seeing a lot of php cycles looking for  & not finding files & so I'm looking at where these might be
addressable.

In this case we can expect that every api call every directory on the path will be checked for the
existing of the relavent file - e.g CustomGroup.php. We really don't need to check for deprecation updates that often.