Fix js error on choosing membership type
I'm seeing a js error when I select a membership type in the back end form. It seems I have recurProcessors assigned
& they have an onChange that calls
buildAutoRenew - which is non existant if the action is not 2. It seems to me that the code is simplest when functions
can be relied on to exist and aree called conditionally - this makes that change, also ensuring 'something' is
assigned for recurProcessors to prevent the error when it's not defined that probably let to the fn being encased in an if

I can't spot any obvious regression cause & perhaps there is something in my config but this might be a regression