Remove unnecessary, and possibly incorrect query from email update
When updating an email the comments pretty clearly say to update bulkmail if it is 1 but the current if clause
picks up anything other than 'null' - which is invalid for a boolean field. This creates an additional query which has some
locking risk being in a high volume place. In addition it seems like it could be a bug in some cases