Remove unnecessary query when updating org, if no name udpate.
This removes an unnecessary query that can be locking at high volume.

The updateCurrentEmployer function sets the organization_name for any related contacts. We don't need to
do this if we are creating the organization for the first time, or updating the organization but not setting (passing in)
organization_name. We can check & avoid in php.

Also we don't need to check for an existing employer_id unless we are dealing with an Individual record. We can
skip that query for Orgs & Households.

These are cheap queries but at high volume can contribute to locks