hook_angularModules - Accept 'partialsCallback' for dynamic partials
This is similar to the antecedent commit which used a field 'snippets' that
contained a list of HTML files and their content. However, this
uses a callback function instead of providing the full HTML content.

This makes it safer to work with the Angular metadata without needing to
load the content of all partials into memory.