dev/event#25 Fix regression on billing name overwrite
https://lab.civicrm.org/dev/event/issues/25 was caused by https://github.com/civicrm/civicrm-core/commit/b1b7f4e35bbe566d9efccac1055d311245420974
where it started overwriting the name in that function.

In digging in the code I found that it seems that as long as we call this function before the processor all is
good so we should call right before the processor rather than broader contamination.

Also note ideally the settings bag would be used.