dev/membership#21 fix regression on membership handling
This turns https://lab.civicrm.org/dev/membership/issues/21#note_29439 into a patch. As new code it seems to
have triggered an underlying issue. I also have https://github.com/civicrm/civicrm-core/pull/16139 open but
would be happy to see this hit the rc & the other have a full rc cycle.

The logic for this change makes sense as described & it is trivial. This has been a very hard bug
to reproduce so I'm unable to give full steps