Remove unindexed join from getACLRoles query
Replaces the indexed join with a  WHERE  on appropriate entity_id values.

I'm not a fan of CRM_Core_OptionGroup::values('acl_role') but it's what is used on the form that saves this.

The form adds in two pseudovalues but they should stay out to respect the INNER join