Remove unused and largely unreachable code.
The removed lines can only be reached if
a) CMS is Joomla or
b) CMS is Drupal and NULL is not equal to NULL

This seemed a little suspicious - especially since the merge screen still does display the value that should
be but isn't calculated here. After digging I found it is duplicated here
https://github.com/civicrm/civicrm-core/blob/bc77d7c0798081f999e6eca92d75fa874fd04e74/CRM/Contact/Form/Merge.php#L115
and that is the value used in the tpl here
https://github.com/civicrm/civicrm-core/blob/1188c7a8512382b1cb588b6a1ea4b13aa9587bc9/templates/CRM/Contact/Form/Merge.tpl#L217

so these lines can be safely removed as redundant. I also signposted where I think the next cleanups should happen on this
with code comments