[REF] Fix CRM_Core_BAO_UFGroup::createUFJoin to not receive by reference.
This function is called from one place. It gets passed a params variable that is 're-used' so this
change makes it clear what actually needs to be passed on. We have an anti-pattern in our code whereby params get
passed all sorts of places with it being confusing which ones actually need to receive what