[REF] Remove apparent copy & paste code.
In code familiarisation to tackle dev#core/1217 I looked at this var & concluded that the var is set in 9
places but in 8 of those it is never used. In the last place (CRM_Contact_Form_Search) it is used to
determine whether to build the pre-next cache. I swapped it over to just doing a getButton without any
change - I stepped through in advanced search, search builder, custom search, basic search it still enters the cache lines of code. In Contribution, event search it doesn't (before & after).

Part of just makingg code semi-readable