[REF] Minor cleanup on  property.
This extracts the few lines of code that originally set this property and later on removes lines of code that
set it again. The removed lines of code are at the end of the submit function. The next fn
getStatusMsg doesn't use  & on return to the postProcess there are just a few calls to getButtonName
that reference id but not _id. So, at this point the variable is being set only to increase confusion