Do not enable core payment processor types that we believe likely don't work on new installs
There are a bunch of payment processors that ship with core for historical reasons that we feel somewhat uncomfortable about. We suspect mos don't work any more and don't know how to support them if they do. We are now seeing
regular test fails on enotices on some & I think disabling them on new installs is a reasonable approach.

We could go further & disable them on existing installs with no active instances (I thought we already had for eWay) but
I think that can be a follow up