dev/core#1364 Merge all addresses on export should INCLUDE merging households
This is an alternate to https://github.com/civicrm/civicrm-core/pull/15725 that still achieves the goal in
https://github.com/civicrm/civicrm-core/pull/15725#issuecomment-572408173 of making is so that the
mergeSharedAddress option mergesHouseholds first & then merges sharedAddresses.

In #15725 @monishdeb was fixing the code that differentiated between how master_id was treated
versus otherwise merged addresses. However that code 'broke' a long time ago and the arguably
broken behaviour is locked in by unit tests. Fixing it has not been requested - in this
issue or any other in the past year or 2 when it has been broken so I opted to remove the broken
code entirely & to open up the next query such that it does not exclude household addresses.
This didn't break any tests (woot) and also allowed me to enable the test rows @monishdeb wrote
to test this issue