Move credit_note_definition on contribution settings form to metadata.
This moves the code that allows settings to be added to forms by metadata from the Generic setting form to the setting trait.

This means many more forms will support injecting settings by metadata. Specifically we switch across the Contribution preferences form,
making credit_note_prefix a metadata-added field. This is preliminary to moving that to an extension & allows us to inject it
in by simply defining the metadata in a hook