Add require_once to api3TestTrait
I used this trait from an extension but it was used with apiv4 before apiv3 (it supports both).

I got an error because utils.php is not included by this file. It 'freeloads' off the fact apiv3 initialises
the file - but in fact the file that includes a function requiring a function
should have the require once - reduces risk & cognitive load