dev/financial#84 Remove sequential credit notes from core to an extension that ships with core.
https://lab.civicrm.org/dev/financial/issues/84

The code that adds  sequentialcredtit notes is a hangover from pre-Lexim days when people's various needs just got added to core.
It would not have been added to core post Lexim for the following reasons
1) architectural - it assumes a one to one relationship between contributions & credit notes which is not reflective of our data model
2) performance - the code had significant performance impact on large sites
3) specificity - the code was for a specific localised use case of needing guaranteed sequential credit-notes
4) maintainability - we've had to put in a lot of work to get the code down to the few lines being removed. The goal is not
to allow organisations to transfer maintenance burden onto the community unless there is a clear rationale.

Medium term the intent is that this extension will not ship with core & will simply be available from the extensions
directory as any other extension. In order to reach this point we need a transitional approach. In the first instance sites
should need to take no action at all in order to experience no change of behaviour. In order to achieve this
we continue to ship this functionality with core as an extension. We enable the extension on upgrade (but on new
installs it is opt-in).

At this point sites who do not want this functionality can disable the extension.

The next step will be to get people who want the extension to install it to their extensions directory
so we can remove this from core later. This will involve us fixing the extension load code to prefer the extension
directory over core if the extension is in both places & adding checks & messaging to encourage people to
install the extension if they continue to want it.

At some point we will remove it from the core tarball. Ongoing maintenance will be provided by whoever wants to maintain  it - if
someone  wants to be added as a maintainer of the extension repo the answer is yes.