Revert "Remove unreachable code lines."
In the course of https://github.com/civicrm/civicrm-core/pull/16462
I concluded that there might be (untested scenarios) where this would still be hit.

In #16462 I adopted a cleaner (tested) approach. However, I think it's unlikely that
will be merged before 5.23 is cut so I propose reverting this change & re-reverting it
once the rc is cut as a more conservative approach. We can focus on getting #16462
wrapped up for 5.24

This reverts commit d7943ca7f7107c1466e687e876bddaf0f1998d99.